+++ Assets/Scripts/GameStateManager/GameStateMachine.cs
@@ -5,9 +5,10 @@ using UnityEngine;
 namespace GameStateManager
 {
     public class GameStateMachine : MonoBehaviour
-    {
-        // References
+    {      // References
         public InputManager _input { get; private set; }
+        public PlayerStatsManager _playerStats { get; private set; }
+
 
         // States
         public readonly GamePlayState PlayState = new();
@@ -15,13 +16,17 @@ namespace GameStateManager
         public readonly GameDeathState DeathState = new();
         public readonly GameLoadingState LoadingState = new();
         public readonly GameMainMenuState MainMenuState = new();
-        
+
         // Components
         public Rigidbody Rb;
         private BaseGameState _currentState;
 
         public void Awake()
         {
+            if (_playerStats == null)
+            {
+                _playerStats = FindObjectOfType<PlayerStatsManager>(); //To Be Reworked When Menu is Added
+            }
             if (_input is null)
             {
                 _input = FindObjectOfType<InputManager>();
@@ -42,7 +47,7 @@ namespace GameStateManager
 
         public void Update()
         {
-            _currentState.OnUpdate();   
+            _currentState.OnUpdate();
         }
 
         public void SwitchState(BaseGameState state)