+++ Assets/Scripts/GameStateManager/PlayerStatsManager.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerStatsManager : MonoBehaviour, IDamageable
+{
+
+ [SerializeField] private float Health = 20;
+
+    void IDamageable.DamageHealth(int damage)
+    {
+
+        Health -= damage;
+    }
+}