+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v1.0.1
+
+## Added
+
+ - Added placeholder rock
+ - Added player stats script
+ - Added interface for damage
+ - Setup ChangeLog
+
+## Updated
+ - Update gamestatemachine to reference PlayerStatsManager
+ 
 ## v1.0.0
 
 ## Added